Move prepare_context to a Rails concern
